% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecter.R
\name{calc_features}
\alias{calc_features}
\title{Compute metrics describing relationship among peptides of the same protein}
\usage{
calc_features(input_dt)
}
\arguments{
\item{input_dt}{data table or data frame in wide representation. The data typically 
contains \code{"PeptideIon"}, \code{"ProteinName"} and sample names in columns and 
measurements of each peptide or precursor ions in rows.}
}
\value{
data.table data.frame containing feature statistics
}
\description{
Compute metrics describing relationship among peptides of the same protein
}
