% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{count_pairwise_number}
\alias{count_pairwise_number}
\title{Number of intersection betwen numeric groups}
\usage{
count_pairwise_number(input_vector_a, input_vector_b)
}
\arguments{
\item{input_vector_a}{vectors containing a sequence of numbers}

\item{input_vector_b}{vectors containing a sequence of numbers}
}
\description{
Number of intersection betwen numeric groups
}
