% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{cv_na}
\alias{cv_na}
\title{Computes coefficient of variation (CV)}
\usage{
cv_na(x)
}
\arguments{
\item{x}{a vector (atomic or list) or an `expression` object.}
}
\value{
a numeric value
}
\description{
A function to compute coefficient of variance (CV) of a numeric vector 
after exclusion of missing values
}
\examples{
vector <- c(1, 2, 4, NA, 2, 5, 6)
cv_na(vector)
}
