% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{dia_guidance}
\alias{dia_guidance}
\title{wrapper function
example
 dia_guidance(data= "S:/SWATH-guidance/feature_alignment.csv", 
                sample_annotation="S:/SWATH-guidance/sample_annotation", 
                level="PeptideIon")}
\usage{
dia_guidance(data, sample_annotation, level)
}
\description{
wrapper function
example
 dia_guidance(data= "S:/SWATH-guidance/feature_alignment.csv", 
                sample_annotation="S:/SWATH-guidance/sample_annotation", 
                level="PeptideIon")
}
