% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Importer.R
\name{import_openswath}
\alias{import_openswath}
\title{import openSWATH output data with sample annotation}
\usage{
import_openswath(search_results, bool.removeDecoy = T, level = "PeptideIon",
  sample_annotation = NULL, remove_prefixInFileName = FALSE)
}
\arguments{
\item{search_results}{A data frame containing the SWATH-MS data. This data typically
contains peptide precursors in each row with corresponding \code{ProteinName}, 
\code{Intensity}, \code{RT}, \code{Score} and etc in columns. The data can be loaded
from local directory.}

\item{bool.removeDecoy}{boolean value (\code{TRUE} or \code{FALSE}) determining 
if the decoy peptides should be removed from the imported data}

\item{level}{the protein-level of imported data. The options include 
\code{"PeptideIon"} (by default), \code{"Transition"}, \code{"Peptide"} or 
\code{"PeptideWithMod"}}

\item{sample_annotation}{data matrix with \code{SampleName}, biological covariates 
(biological replicates) and technical covariates (technical replicates, batches, etc)}

\item{remove_prefixInFileName}{boolean value (\code{TRUE} or \code{FALSE}) whether to 
remove unnecessary prefix from euler portal file name. For example, a typical 
filename will look like \code{"/scratch/71239421.tmpdir/xuep_J180621_SW_3.mzXML.gz"} 
and \code{remove_prefixInFileName = TRUE} will result in \code{"xuep_J180621_SW_3.mzXML.gz"}}
}
\value{
data.table data.frame containing SWATH-MS data and sample annotation
}
\description{
import openSWATH output data with sample annotation
}
