% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Importer.R
\name{import_openswath_matrix_fromEulerPortal}
\alias{import_openswath_matrix_fromEulerPortal}
\title{Import openSWATH data table from Eular portal}
\usage{
import_openswath_matrix_fromEulerPortal(search_results,
  sample_annotation = NULL)
}
\arguments{
\item{search_results}{A data frame containing the SWATH-MS data. This data typically
contains peptide ions in each row with corresponding \code{ProteinName}, 
\code{Intensity}, \code{RT}, \code{Score} and etc in columns. The data can be loaded
from local directory.}

\item{sample_annotation}{data matrix with \code{SampleName}, biological covariates 
(biological replicates) and technical covariates (technical replicates, batches, etc)}
}
\value{
data.table data containing raw openSWATH matrix
}
\description{
A function to import a full openSWATH data table, 
typically pre-processed or pre-filtered for downstream analysis. 
The imported data table contains columns denoting peptide name, corresponding 
protein name, intensity, retention time and score of each peptide.
}
\examples{
all_peptideIons <- import_openswath_matrix_fromEulerPortal(
search_results="data/YS_SWATH_data", 
sample_annotation="data/YS_sample_annotation")

}
