% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Importer.R
\name{import_spectronaut_matrix}
\alias{import_spectronaut_matrix}
\title{Import Spectronaut data table}
\usage{
import_spectronaut_matrix(search_results, sample_annotation = NULL)
}
\arguments{
\item{search_results}{A data frame containing the SWATH-MS data. This data typically
contains peptide ions in each row with corresponding \code{ProteinName}, 
\code{Intensity}, \code{RT}, \code{Score} and etc in columns. The data can be loaded
from local directory.}

\item{sample_annotation}{data matrix with \code{SampleName}, biological covariates 
(biological replicates) and technical covariates (technical replicates, batches, etc)}
}
\description{
A function to import a full Spectronaut data table, typically 
pre-processed or pre-filtered for downstream analysis. The imported data 
table contains columns denoting peptide name, corresponding protein name, 
intensity, retention time and score of each peptide. From the Spectronaut 
output, this function takes PEP Quantity value as Intensity and EG q-value 
as Score.
}
\examples{
peptideIons <- import_spectronaut_matrix(search_results= "data/QGS_SWATH_data", 
sample_annotation="data/QGS_sample_annotation")

}
