% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizer.R
\name{keep_proteotypic_only}
\alias{keep_proteotypic_only}
\title{Filter for proteotypic peptides}
\usage{
keep_proteotypic_only(input_dt)
}
\arguments{
\item{input_dt}{data table or data frame in wide representation. The data typically 
contains \code{"PeptideIon"}, \code{"ProteinName"} and sample names in columns and 
measurements of each peptide or precursor ions in rows. 

The input data should contain \code{ProteinName} column which denotes an 
unique identifier for protein or proteingroup that the peptide maps to. 
Proteotypic peptides should be indicated by 1/ to be recognized 
by \code{keep_proteotypic_only()}}
}
\value{
data.table data.frame containing only proteotypic peptides
}
\description{
A function to filter for proteotypic peptides indicated as 1/ 
under \code{ProteinName}. For any given protein, only a few proteotypic peptides 
are uniquely and consistently identified and these represent peptides 
most confidently observed in the current MS-based proteomics workflow.
}
\examples{
proteotypic_peptide_matrix <- keep_proteotypic_only(all_peptideIons)

}
