% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{l_count}
\alias{l_count}
\title{Count TRUE values}
\usage{
l_count(x)
}
\arguments{
\item{x}{a vector (atomic or list) in TRUE or FALSE}
}
\value{
a numeric value
}
\description{
A function to count the number of TRUE values in a vector
}
\examples{
vector <- c(TRUE, TRUE, FALSE, TRUE)
l_count(vector)

}
