% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizer.R
\name{long2wide}
\alias{long2wide}
\title{Convert from long to wide data}
\usage{
long2wide(input_dt, global_level = "PeptideIon")
}
\arguments{
\item{input_dt}{a data table or data frame in a long format. The data typically 
contains peptide ID (e.g. \code{PeptideIon}), sample ID (e.g. \code{SampleName}) and 
measurements (e.g. \code{Intensity}, \code{RT} and \code{Score}).}

\item{global_level}{the protein-level of imported data. The options include 
\code{"PeptideIon"} (by default), \code{"Transition"}, \code{"Peptide"} or 
\code{"PeptideWithMod"}. The \code{global_value} specifies format of input data
and thus selects suitable columns for conversion into wide format.}
}
\value{
data.table data.frame in a wide format representation containing 
\code{"PeptideIon"}, \code{"ProteinName"} and sample names in columns and 
measurements of each peptide or precursor ions in rows.
}
\description{
A function to convert from long to wide representation. 
A typical input data table contains columns denoting name of a peptide, 
its corresponding protein name, sample name, intensity, retention time and 
width of spectral peaks. The \code{long2wide()} converts this long format data 
into wide representation having intensity, score, RT and width of each sample 
in each column. Each row will represent a unique peptide.
}
\examples{
peptideIons <- import_openswath(search_results= "data/QGS_SWATH_data", 
sample_annotation="data/QGS_sample_annotation", 
level="PeptideIon")
all_peptideIons <- long2wide(peptideIons)

}
