% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{mean_na}
\alias{mean_na}
\title{Computes average excluding missing values}
\usage{
mean_na(x)
}
\arguments{
\item{x}{a vector (atomic or list) or an `expression` object.}
}
\value{
a numeric value
}
\description{
Computes average excluding missing values
}
