% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{median_na}
\alias{median_na}
\title{Computes median}
\usage{
median_na(x)
}
\arguments{
\item{x}{a vector (atomic or list) or an `expression` object.}
}
\value{
a numeric value
}
\description{
A function to compute median of a numeric vector 
after exclusion of missing values
}
\examples{
vector <- c(1, 2, 4, NA, 2, 5, 6)
median_na(vector)
}
