% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizer.R
\name{merge_replicates}
\alias{merge_replicates}
\title{Merge biological replicates}
\usage{
merge_replicates(wide, sample_annotation = NULL,
  bool_NA_means_requant = FALSE, averageFun = "mean")
}
\arguments{
\item{wide}{data table or data frame in wide representation. The data typically 
contains \code{PeptideIon}, \code{ProteinName} and sample names in columns and 
measurements of each peptide or precursor ions in rows.}

\item{sample_annotation}{data matrix with \code{SampleName}, biological covariates 
(biological replicates) and technical covariates (technical replicates, batches, etc)}

\item{bool_NA_means_requant}{boolean value (\code{TRUE} or \code{FALSE}) determining if 
the missing values correspond to requants. \code{bool_NA_means_requant = TRUE} will 
use the number of requant values (m_score = 2) as the number of NAs.}

\item{averageFun}{method to compute mean peptide or precursor ion intensity 
of biological replicates. Options include \code{"mean"} and \code{"median"}.}
}
\value{
data.table data.frame containing statistics of biological replicates, such 
as \code{mean_intensity_A}, \code{cv_intensity_A} and \code{numNA_intensity_A}.
}
\description{
Merge biological replicates
}
\examples{
all_peptideIons_normalized <- normalize_data(all_peptideIons, replaceNA="keep", 
normalization="none")
cons_peptideIons <- merge_replicates(all_peptideIons_normalized, anno)

}
