% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizer.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize data}
\usage{
normalize_data(input_dt, replaceNA = "keep",
  normalization = "mediancenter")
}
\arguments{
\item{input_dt}{data table or data frame in wide representation. The data typically 
contains \code{"PeptideIon"}, \code{"ProteinName"} and sample names in columns and 
measurements of each peptide or precursor ions in rows.}

\item{replaceNA}{whether to treat missing values. The options include to \code{"remove"}, 
\code{"keep"}, replace them with \code{"zero"}, or minimum intensity (\code{"min_intensity"})}

\item{normalization}{different methods of normalization. The options include 
median-centering (\code{"mediancenter"}), quantile normalization (\code{"quantile"},
and normalized based on total ion current (\code{"TIC"}) and indexed retention 
time (iRT) standards \code{"iRT"}. The default is \code{"mediancenter"} and denote 
\code{"none"} if normalization is not necessary.}
}
\value{
data.table data.frame containing normalized measurement data and a column
denoting number of peptides corresponding to the same protein
}
\description{
A function to normalize peptide/fragment-level intensity across 
different samples. It provides several normalization methods: 
\code{“mediancenter”} as default method to equalize median of intensity distribution, 
\code{“quantile”} to adjust distributions to have the same statistical properties 
i.e. quantiles, \code{“TIC”} to normalize mass spectra based on total ion 
current (TIC) assuming that the same amount of protein was injected for each 
MS run, and lastly \code{“iRT”} to normalize intensities based on injected 
indexed retention time (iRT) standards. 

In addition to the normalization functionality, this function also counts 
the number of peptides corresponding to the same protein and outputs in 
column \code{numPerProt}.
}
\examples{
peptideIon_n <- normalize_data(peptideIon_st, replaceNA="keep", normalization="none")

}
