% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{pairwise_length_matrix}
\alias{pairwise_length_matrix}
\title{Number of intersection within a matrix}
\usage{
pairwise_length_matrix(input_matrix)
}
\arguments{
\item{input_matrix}{a matrix containing numeric values}
}
\description{
A function to calculate the number of intersecting values 
between rows of input matrix
}
