% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{pairwise_length_vector}
\alias{pairwise_length_vector}
\title{Number of intersection}
\usage{
pairwise_length_vector(input_a, input_b)
}
\arguments{
\item{input_a}{vectors, data frames, or ps objects
containing a sequence of elements (conceptually).}

\item{input_b}{vectors, data frames, or ps objects
containing a sequence of elements (conceptually).}
}
\description{
A function to calculate the number of intersecting 
values between two vectors, excluding missing values
}
\examples{
vector1 <- c(1, 2, 4, NA, 2, 5, 6)
vector2 <- c(1, 2, 4, NA, 6, 7, 8)
pairwise_length_vector(vector1, vector2)

}
