% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticalTest.R
\name{perform_anova}
\alias{perform_anova}
\title{Perform ANOVA}
\usage{
perform_anova(input_dt, sample_annotation)
}
\arguments{
\item{input_dt}{@param input_dt data table or data frame in wide representation. The data typically 
contains \code{"PeptideIon"}, \code{"ProteinName"} and sample names in columns and 
protein measurements in rows. The input data is usually an output from 
\code{pept2prot()} and \code{merge_replicates()}}

\item{sample_annotation}{data matrix with \code{SampleName}, biological covariates 
(biological replicates) and technical covariates (technical replicates, batches, etc)}
}
\value{
data.table data.frame containing statistics of parametric and 
non-parametric anova results
}
\description{
A function to perform Analysis of variance (ANOVA) test to identify
differentially expressed proteins between different sample groups.
}
\examples{
anova_results <- perform_anova(protein_Filtered_top3_sum_ImputedWeighted)

}
