% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statisticalTest.R
\name{perform_modt_tests}
\alias{perform_modt_tests}
\title{Performed modified t.test}
\usage{
perform_modt_tests(input_dt, sample_annotation, input_bool_paired = FALSE,
  input_mtc_method = "BH")
}
\arguments{
\item{input_dt}{data table or data frame in wide representation. The data typically 
contains \code{"PeptideIon"}, \code{"ProteinName"} and sample names in columns and 
protein measurements in rows. The input data is usually an output from 
\code{pept2prot()} and \code{merge_replicates()}}

\item{sample_annotation}{data matrix with \code{SampleName}, biological covariates 
(biological replicates) and technical covariates (technical replicates, batches, etc)}

\item{input_bool_paired}{a logical indicating whether a paired test is performed}

\item{input_mtc_method}{multiple testing correction method. Options include
\code{"holm"}, \code{"hochberg"}, \code{"hommel"}, \code{"bonferroni"}, 
\code{"BH"}, \code{"BY"}, \code{"fdr"} and \code{"none"}. Refer to 
\code{p.adjust()} for details.}
}
\description{
Performed modified t.test
}
