% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecter.R
\name{perform_selection}
\alias{perform_selection}
\title{Compute probability of being representative peptide}
\usage{
perform_selection(input_dt)
}
\arguments{
\item{input_dt}{data table or data frame in wide representation. The data typically 
contains \code{"PeptideIon"}, \code{"ProteinName"} and sample names in columns and 
measurements of each peptide or precursor ions in rows. 

Additionally, the data table includes metrics computed from \code{calc_features()} 
such as \code{scaled_mean_intensity_all}, \code{scaled_cv_intensity_all} 
\code{scaled_numNA_intensity_all} and etc which are utilized to compute 
a posterior probability of being the representative peptides.}
}
\value{
data.table data.frame containing posterior probability.
}
\description{
A function to compute the probability of being representative 
peptide. It utilizes \code{predict()} from \code{stats} to make prediction
based on a linear discriminant analysis (LDA) model, typically generated 
from a quantitative gold standard dataset. The function outputs the posterior
probability that describes the likelihood of the peptide intensity being
representative of the intensity of its corresponding protein.
}
\examples{
global_level = "PeptideIon"
d_feature <- calc_features(peptideIon_st)
d_select <- perform_selection(d_feature)

}
