% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizer.R
\name{plot_a_heatmap_include_prob_update}
\alias{plot_a_heatmap_include_prob_update}
\title{Visualization of protein abundance}
\usage{
plot_a_heatmap_include_prob_update(case, cutoff_prob = 0.3)
}
\description{
Visualization of protein abundance
}
