% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizer.R
\name{plot_bar_intensity_n_probability}
\alias{plot_bar_intensity_n_probability}
\title{Visualize intensity and posterior probability in barplots}
\usage{
plot_bar_intensity_n_probability(case, cutoff_prob = 0.3,
  plot_intensity = T, plot_prob = T)
}
\arguments{
\item{case}{data table or data frame in wide representation. The data typically 
contains \code{"PeptideIon"}, \code{"ProteinName"} and sample names in columns and 
measurements of each peptide or precursor ions in rows. 

To visualize profile of peptides corresponding to the same protein, 
peptides corresponding to one protein are pre-filtered for clearer 
visualizations.}

\item{cutoff_prob}{a numeric value denoting posterior probability threshold
to keep or remove peptides.}

\item{plot_intensity}{a logical indicating whether to plot intensity barplots}

\item{plot_prob}{a logical indicating whether to plot posterior probability
barplots}
}
\description{
Visualize intensity and posterior probability in barplots
}
\examples{
peptideIons_features <- calc_features(all_peptideIons)
peptideIons_features_select <- perform_selection(peptideIons_features)

prot_name <- c("1/O75976")
test_prot <- test[test$ProteinName==prot_name, ]
plot_bar_intensity_n_probability(test_prot)

}
