% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizer.R
\name{plot_cor_heatmap}
\alias{plot_cor_heatmap}
\title{Visualize correlation heatmap}
\usage{
plot_cor_heatmap(case, cutoff_prob = 0.3)
}
\arguments{
\item{case}{data table or data frame in wide representation. The data typically 
contains \code{"PeptideIon"}, \code{"ProteinName"} and sample names in columns and 
measurements of each peptide or precursor ions in rows. 

To visualize profile of peptides corresponding to the same protein, 
recommend to prepare a data table filtered on peptides 
corresponding to a protein of interest}

\item{cutoff_prob}{a numeric value denoting posterior probability threshold
to keep or remove peptides.}
}
\description{
A function to visualize heatmap representing correlation between 
different peptides corresponding to either the same or different proteins.
}
\examples{
global_level = "PeptideIon"
d_feature <- calc_features(peptideIon_st)
d_feature_select <- perform_selection(d_feature)

prot_name <- c("1/sp|O00429|DNM1L_HUMAN")
test_prot <- d_feature_select[d_feature_select$ProteinName==prot_name, ]
p <- plot_cor_heatmap(test_prot)

}
