% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizer.R
\name{plot_density}
\alias{plot_density}
\title{Plot density distribution of feature}
\usage{
plot_density(data, feature = "feature_mean_intensity_all", fill = NULL,
  font.size = 12, title = feature)
}
\description{
Plot density distribution of feature
}
