% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizer.R
\name{plot_peptide_intensity}
\alias{plot_peptide_intensity}
\title{Visualize peptide intensity}
\usage{
plot_peptide_intensity(case, cutoff_prob = 0.3)
}
\arguments{
\item{case}{data table or data frame in wide representation. The data typically 
contains \code{"PeptideIon"}, \code{"ProteinName"} and sample names in columns and 
measurements of each peptide or precursor ions in rows. 

To visualize profile of peptides corresponding to the same protein, 
peptides corresponding to one protein are pre-filtered for clearer 
visualizations.}

\item{cutoff_prob}{a numeric value denoting posterior probability threshold
to keep or remove peptides.}
}
\description{
Visualize peptide intensity
}
