% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizer.R
\name{plot_protein_profile}
\alias{plot_protein_profile}
\title{Visualization of peptides}
\usage{
plot_protein_profile(case, cutoff_prob = 0.3)
}
\arguments{
\item{case}{}

\item{cutoff_prob}{}
}
\description{
Visualization of peptides
}
