% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{replace_inf}
\alias{replace_inf}
\title{replace infinity values by NA}
\usage{
replace_inf(input_dt)
}
\arguments{
\item{input_dt}{a data table, data frame or matrix with numeric values}
}
\value{
data.table
}
\description{
replace infinity values by NA
}
