% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{sd_na}
\alias{sd_na}
\title{Computes standard deviation}
\usage{
sd_na(x)
}
\arguments{
\item{x}{a vector (atomic or list) or an `expression` object.}
}
\value{
a numeric value
}
\description{
A function to compute standard deviation of a numeric vector 
after exclusion of missing values
}
\examples{
vector <- c(1, 2, 4, NA, 2, 5, 6)
sd_na(vector)
}
