% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{sum_na}
\alias{sum_na}
\title{Computes sum}
\usage{
sum_na(x)
}
\arguments{
\item{x}{a vector (atomic or list) or an `expression` object.}
}
\value{
a numeric value
}
\description{
A function to compute sum of a numeric vector 
after exclusion of missing values
}
\examples{
vector <- c(1, 2, 4, NA, 2, 5, 6)
sum_na(vector)
}
