% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{u_count}
\alias{u_count}
\title{Count number of unique values}
\usage{
u_count(x)
}
\arguments{
\item{x}{a vector (atomic or list) or an `expression` object.}
}
\value{
a numeric value
}
\description{
A function to count the number of unique values in a numeric 
or character vector
}
\examples{
vector <- c("c", "d", "v", "d", "t", "c", "a")
u_count(vector)
}
