% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecter.R
\name{calc_features}
\alias{calc_features}
\title{Compute metrics describing relationship among peptides of the same protein}
\usage{
calc_features(input_dt)
}
\arguments{
\item{input_dt}{data table or data frame in wide representation. The data typically 
contains \code{"PeptideIon"}, \code{"ProteinName"} and sample names in columns and 
measurements of each peptide or precursor ions in rows.}
}
\value{
data.table data.frame containing feature statistics
}
\description{
A function to compute feature i.e. peptide statistics including 
average intensity, coefficient of variance (CV), number of missing values, 
average score, standard deviation, median Pearson correlation coefficient (PCC), 
median Spearman correlation coefficient (SCC) and median absolute deviation (MAD). 
The function also outputs a set of scaled statistics, where numeric values are 
centered by subtracting the mean and scaled by dividing standard deviation.
}
\examples{
 d_feature <- calc_features(peptideIon_st)

}
