% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{cv_na}
\alias{cv_na}
\title{Computes coefficient of variation (CV) excluding missing values}
\usage{
cv_na(x)
}
\arguments{
\item{x}{a vector (atomic or list) or an `expression` object.}
}
\description{
Computes coefficient of variation (CV) excluding missing values
}
