% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selecter.R
\name{get_lda_model}
\alias{get_lda_model}
\title{Generate linear discriminant analysis (LDA) model}
\usage{
get_lda_model(input_dt, input_features)
}
\arguments{
\item{input_dt}{data table or data frame in wide representation. The data typically 
contains \code{"PeptideIon"}, \code{"ProteinName"} and sample names in columns and 
measurements of each peptide or precursor ions in rows.}

\item{input_features}{a vector of features to be used for training. The examples 
include \code{scaled_mean_intensity_all}, \code{scaled_cv_intensity_all}, 
\code{scaled_numNA_intensity_all}, \code{scaled_averaged_score_all}, 
\code{scaled_median_PCC}, \code{scaled_sd_width_all} and \code{label}.}
}
\description{
Generate linear discriminant analysis (LDA) model
}
