% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Importer.R
\name{import_openswath}
\alias{import_openswath}
\title{cluster the data matrix to visually inspect which confounder dominates}
\usage{
import_openswath(search_results, bool.removeDecoy = T,
  level = "PeptideIon", sample_annotation = NULL,
  remove_prefixInFileName = FALSE)
}
\arguments{
\item{search_results}{}

\item{bool.removeDecoy}{whether to remove docoy peptides in importing data}

\item{level}{}

\item{sample_annotation}{data matrix with: \enumerate{ \item \code{sample_id_col}
(this can be repeated as row names) \item biological covariates \item
technical covariates (batches etc) }}

\item{remove_prefixInFileName}{}

\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. in most
function, it is assumed that this is the log transformed version of the
original data}

\item{color_df}{data frame of colors, as created by \code{sample_annotation_to_colors}}

\item{data_matrix}{features (in rows) vs samples (in columns) matrix, with
feature IDs in rownames and file/sample names as colnames. Usually the log
transformed version of the original data}

\item{distance}{distance metric used for clustering}

\item{agglomeration}{agglomeration methods as used by \code{hclust}}

\item{label_samples}{if \code{TRUE} sample IDs (column names of 
\code{data_matrix}) will be printed}

\item{label_font}{size of the font. Is active if \code{label_samples} is 
\code{TRUE}, ignored otherwise}

\item{plot_title}{Title of the plot (usually, processing step + representation
level (fragments, transitions, proteins))}

\item{...}{other parameters of \code{plotDendroAndColors} from \code{WGCNA} package}

\item{sample_annotation}{}
}
\value{
No return
}
\description{
cluster the data matrix to visually inspect which confounder dominates
}
\examples{
color_scheme <- sample_annotation_to_colors (example_sample_annotation, 
factor_columns = c('MS_batch','EarTag', "Strain", "Diet", "digestion_batch", "Sex"),
not_factor_columns = 'DateTime',
numeric_columns = c('order'))

color_annotation <- color_scheme$color_df

plot_hierarchical_clustering(example_proteome_matrix, color_annotation,  
distance = "euclidean", agglomeration = 'complete',
label_samples = FALSE)

}
\seealso{
\code{\link[stats]{hclust}}, \code{\link{sample_annotation_to_colors}},
  \code{\link[WGCNA]{plotDendroAndColors}}
import openSWATH output data with sample annotation
}
