% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantification.R
\name{impute_missing_values}
\alias{impute_missing_values}
\title{Impute missing values}
\usage{
impute_missing_values(input_dt, input_index)
}
\arguments{
\item{input_dt}{data table or data frame in wide representation. The data typically 
contains \code{PeptideIon}, \code{ProteinName} and sample names in columns and 
measurements of each peptide or precursor ions in rows.}

\item{input_index}{a integer vector denoting numeric columns to apply imputation}
}
\value{
data.table data.frame
}
\description{
A function to impute for missing value by fitting data to a 
uniform distribution between minimum $0.1 \times CV$ and minimum 
$1.1 \times CV$ and infer the intensity of missing values.
}
\examples{
peptideIons_features <- calc_features(all_peptideIons)
peptideIons_features_select <- calc_features(peptideIons_features)
Imputated <- impute_missing_values(test_yesFiltered, c(3:17))

}
