% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{l_count}
\alias{l_count}
\title{Count number of values in a vector}
\usage{
l_count(x)
}
\arguments{
\item{x}{a vector (atomic or list) or an `expression` object.}
}
\value{
a numeric value
}
\description{
Count number of values in a vector
}
