% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizer.R
\name{long2wide}
\alias{long2wide}
\title{Convert from long to wide representation}
\usage{
long2wide(input_dt, global_level = "PeptideIon")
}
\arguments{
\item{input_dt}{a data table or data frame in a long format. The data typically 
contains peptide ID (e.g. \code{PeptideIon}), sample ID (e.g. \code{SampleName}) and 
measurements (e.g. \code{Intensity}, \code{RT} and \code{Score}).}

\item{global_level}{the protein-level of imported data. The options include 
\code{"PeptideIon"} (by default), \code{"Transition"}, \code{"Peptide"} or 
\code{"PeptideWithMod"}. The \code{global_value} specifies format of input data
and thus selects suitable columns for conversion into wide format.}
}
\value{
data.table data.frame in a wide format representation containing 
\code{"PeptideIon"}, \code{"ProteinName"} and sample names in columns and 
measurements of each peptide or precursor ions in rows.
}
\description{
Convert from long to wide representation
}
