% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{median_na}
\alias{median_na}
\title{Computes median excluding missing values}
\usage{
median_na(x)
}
\arguments{
\item{x}{a vector (atomic or list) or an `expression` object.}
}
\value{
a numeric value
}
\description{
Computes median excluding missing values
}
