% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalizer.R
\name{normalize_data}
\alias{normalize_data}
\title{Normalize data}
\usage{
normalize_data(input_dt, replaceNA = "keep",
  normalization = "mediancenter")
}
\arguments{
\item{input_dt}{data table or data frame in wide representation. The data typically 
contains \code{"PeptideIon"}, \code{"ProteinName"} and sample names in columns and 
measurements of each peptide or precursor ions in rows.}

\item{replaceNA}{whether to treat missing values. The options include to \code{"remove"}, 
\code{"keep"}, replace them with \code{"zero"}, or minimum intensity (\code{"min_intensity"})}

\item{normalization}{different methods of normalization. The options include 
median-centering (\code{"mediancenter"}), quantile normalization (\code{"quantile"},
and normalized based on total ion current (\code{"TIC"}) and indexed retention 
time (iRT) standards \code{"iRT"}. The default is \code{"mediancenter"} and denote 
\code{"none"} if normalization is not necessary.}
}
\value{
data.table data.frame containing normalized measurement data
}
\description{
Normalize data
}
