% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{pairwise_length_vector}
\alias{pairwise_length_vector}
\title{Number of intersection between two groups}
\usage{
pairwise_length_vector(input_a, input_b)
}
\arguments{
\item{input_a}{vectors, data frames, or ps objects
containing a sequence of elements (conceptually).}

\item{input_b}{vectors, data frames, or ps objects
containing a sequence of elements (conceptually).}
}
\description{
Number of intersection between two groups
}
