% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantification.R
\name{pept2prot}
\alias{pept2prot}
\title{Infer protein abudnace from peptide measurements}
\usage{
pept2prot(input_dt, input_rank_index = "prob", topN = 3, aggfun = "mean",
  bool_weighted_by_prob = TRUE)
}
\arguments{
\item{input_dt}{data table or data frame in wide representation. The data typically 
contains \code{PeptideIon}, \code{ProteinName} and sample names in columns and 
measurements of each peptide or precursor ions in rows}

\item{topN}{number of peptides utilized to infer protein abundance for each protein}

\item{aggfun}{method to aggregate peptide measurements to estimate protein abundance.
Options include \code{"mean"} and \code{"sum"}}

\item{bool_weighted_by_prob}{boolean value (\code{TRUE} or \code{FALSE}) determining 
whether to weight the intensity by the posterior probability of being a representative 
peptide}
}
\description{
Infer protein abudnace from peptide measurements
}
