% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizer.R
\name{plot_bar_intensity_n_probability}
\alias{plot_bar_intensity_n_probability}
\title{Visualize intensity and posterior probability in barplots}
\usage{
plot_bar_intensity_n_probability(case, cutoff_prob = 0.3,
  bool_isPeptideIon = TRUE, plot_intensity = T, plot_prob = T)
}
\arguments{
\item{case}{data table or data frame in wide representation. The data typically 
contains \code{"PeptideIon"}, \code{"ProteinName"} and sample names in columns and 
measurements of each peptide or precursor ions in rows. 

To visualize profile of peptides corresponding to the same protein, 
recommend to prepare a data table filtered on peptides 
corresponding to a protein of interest}

\item{cutoff_prob}{a numeric value denoting posterior probability threshold
to keep or remove peptides.}

\item{plot_intensity}{a logical indicating whether to plot intensity barplots}

\item{plot_prob}{a logical indicating whether to plot posterior probability
barplots}
}
\description{
A function to visualize average peptide intensity and 
posterior probability of being the representative peptide via barplots.
}
\examples{
global_level = "PeptideIon"
d_feature <- calc_features(peptideIon_st)
d_feature_select <- perform_selection(d_feature)

prot_name <- c("1/sp|O00429|DNM1L_HUMAN")
test_prot <- d_feature_select[d_feature_select$ProteinName==prot_name, ]
p <- plot_bar_intensity_n_probability(test_prot)

}
