% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizer.R
\name{plot_density}
\alias{plot_density}
\title{Plot density distribution of peptide features}
\usage{
plot_density(data, feature = "feature_mean_intensity_all", fill = NULL,
  font.size = 12, title = feature)
}
\arguments{
\item{data}{data table or data frame in wide representation. The data typically 
contains \code{PeptideIon}, \code{ProteinName} and sample names in columns and 
measurements of each peptide or precursor ions in rows. The function is 
particularly useful after \code{calc_features()} step to visualize distribution of 
computed feature statistics.}

\item{feature}{a character vector denoting a peptide feature for density 
plots. Examples include \code{"feature_mean_intensity_all"}, 
\code{"feature_cv_intensity_all"}, \code{"scaled_median_PCC"} and etc.}

\item{fill}{a character vector denoting a column name to color by}

\item{font.size}{font size of x and y labels}

\item{title}{title of density plot}
}
\description{
A function to visualize distribution of computed features through 
density plots. These features include average intensity, coefficient of variance 
and standard deviation.
}
\examples{
global_level = "PeptideIon"
d_feature <- calc_features(peptideIon_st)

p <- plot_density(d_feature, feature = "feature_mean_intensity_all")

}
