% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualizer.R
\name{plot_peptide_intensity}
\alias{plot_peptide_intensity}
\title{Visualize peptide intensity}
\usage{
plot_peptide_intensity(case, cutoff_prob = 0.3,
  bool_isPeptideIon = TRUE)
}
\arguments{
\item{case}{data table or data frame in wide representation. The data typically 
contains \code{"PeptideIon"}, \code{"ProteinName"} and sample names in columns and 
measurements of each peptide or precursor ions in rows. For a clear visualization
of peptide profile, recommend to prepare a data table filtered on peptides 
corresponding to a protein of interest}

\item{cutoff_prob}{a numeric value denoting posterior probability threshold
to keep or remove peptides.}
}
\description{
Visualize peptide intensities in different samples by a line graph. 
Lines are colored by differnet peptides and dotted represent removed peptides based on 
LDA algorithm.
}
\examples{
peptideIons_features <- calc_features(all_peptideIons)
peptideIons_features_select <- perform_selection(peptideIons_features)

prot_name <- c("1/O75976")
test_prot <- test[test$ProteinName==prot_name, ]
p <- plot_peptide_intensity(test_prot)

}
