% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guidance.R
\docType{package}
\name{proBatch}
\alias{proBatch}
\alias{proBatch-package}
\title{DIA-guidance: A package to infer protein abundance from DIA/SWATH-MS 
peptide measurements}
\description{
The DIA-guidance package contains functions to normalize peptide/fragment-level intensities, 
estimate protein abundances and conduct statistical tests on protein measurements. Although 
the package has primarily been developed for mass spectrometry proteomics (DIA/SWATH),
it should also be applicable to most proteomics data with minor adaptations.
}
\details{
To learn more about proBatch, start with the vignettes:
\code{browseVignettes(package = "guidance")}
}
