% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{replace_inf}
\alias{replace_inf}
\title{replace infinity with NA}
\usage{
replace_inf(input_dt)
}
\arguments{
\item{input_dt}{data.table or data frame with numeric values}
}
\value{
data.table
}
\description{
replace infinity with NA
}
