% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{sd_na}
\alias{sd_na}
\title{Computes standard deviation excluding missing values}
\usage{
sd_na(x)
}
\arguments{
\item{x}{a vector (atomic or list) or an `expression` object.}
}
\value{
a numeric value
}
\description{
Computes standard deviation excluding missing values
}
