% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic.R
\name{sum_na}
\alias{sum_na}
\title{Computes sum excluding missing values}
\usage{
sum_na(x)
}
\arguments{
\item{x}{a vector (atomic or list) or an `expression` object.}
}
\description{
Computes sum excluding missing values
}
