% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Writer.R
\name{write_protein_table}
\alias{write_protein_table}
\title{export protein table to tsv or csv file}
\usage{
write_protein_table(table, name = "protein_table", format = "tsv")
}
\description{
export protein table to tsv or csv file
}
