% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Writer.R
\name{write_protein_table}
\alias{write_protein_table}
\title{Export data table}
\usage{
write_protein_table(input_dt, file_name = "protein_table",
  format = "tsv", ...)
}
\arguments{
\item{input_dt}{the object to be written, preferably data frame, data table or matrix}

\item{file_name}{name of exported file}

\item{format}{format the object to be exported. Options include \code{"tsv"} and 
\code{"csv"}.}

\item{...}{other parameters of \code{write.table} function}
}
\value{
csv or tsv output
}
\description{
Export data table e.g. protein table to a tsv or csv file.
}
\examples{
\dontrun{write_protein_table(protein_matrix, 
file_name = "example_protein_table", format = "tsv")}

}
